Move button config to it's own file
