+++ src/components/Pagination/styles/Pagination.css.js
@@ -3,7 +3,7 @@ import baseStyles from '../../../styles/resets/baseStyles.css.js'
 import { getColor } from '../../../styles/utilities/color'
 import Button from '../../Button'
 
-import { config } from '../../Button/styles/Button.css'
+import config from '../../Button/styles/Button.config.css.js'
 
 export const PaginationUI = styled('div')`
   ${baseStyles};
